--- a/terraform-aws/main.tf
+++ b/terraform-aws/main.tf
@@ -17,8 +17,8 @@ resource "aws_instance" "web_app" {
   ebs_block_device {
     device_name = "my_data"
     volume_type = "io1"                     # <<<<< Try changing this to gp2 to compare costs
-    volume_size = 1000
-    iops        = 800
+    volume_size = 3000
+    iops        = 1200
   }
 }