--- a/terraform-gcp/main.tf
+++ b/terraform-gcp/main.tf
@@ -20,7 +20,7 @@ resource "google_compute_instance" "instance1" {
 
   guest_accelerator {
     type = "nvidia-tesla-t4" # <<<<< Try changing this to nvidia-tesla-p4 to compare the costs
-    count = 4
+    count = 3
   }
 
   network_interface {