@aliscott 2024-04-10
Ran Plan for 2 projects:

1. dir: `terraform-aws` workspace: `default`
1. dir: `terraform-gcp` workspace: `default`

### 1. dir: `terraform-aws` workspace: `default`
<details><summary>Show Output</summary>

```diff
Terraform used the selected providers to generate the following execution
plan. Resource actions are indicated with the following symbols:
+ create

Terraform will perform the following actions:

  # aws_instance.web_app will be created
+ resource "aws_instance" "web_app" {
      + ami                                  = "ami-674cbc1e"
      + arn                                  = (known after apply)
      + associate_public_ip_address          = (known after apply)
      + availability_zone                    = (known after apply)
      + cpu_core_count                       = (known after apply)
      + cpu_threads_per_core                 = (known after apply)
      + disable_api_stop                     = (known after apply)
      + disable_api_termination              = (known after apply)
      + ebs_optimized                        = (known after apply)
      + get_password_data                    = false
      + host_id                              = (known after apply)
      + host_resource_group_arn              = (known after apply)
      + iam_instance_profile                 = (known after apply)
      + id                                   = (known after apply)
      + instance_initiated_shutdown_behavior = (known after apply)
      + instance_lifecycle                   = (known after apply)
      + instance_state                       = (known after apply)
      + instance_type                        = "m5.4xlarge"
      + ipv6_address_count                   = (known after apply)
      + ipv6_addresses                       = (known after apply)
      + key_name                             = (known after apply)
      + monitoring                           = (known after apply)
      + outpost_arn                          = (known after apply)
      + password_data                        = (known after apply)
      + placement_group                      = (known after apply)
      + placement_partition_number           = (known after apply)
      + primary_network_interface_id         = (known after apply)
      + private_dns                          = (known after apply)
      + private_ip                           = (known after apply)
      + public_dns                           = (known after apply)
      + public_ip                            = (known after apply)
      + secondary_private_ips                = (known after apply)
      + security_groups                      = (known after apply)
      + source_dest_check                    = true
      + spot_instance_request_id             = (known after apply)
      + subnet_id                            = (known after apply)
      + tags_all                             = (known after apply)
      + tenancy                              = (known after apply)
      + user_data                            = (known after apply)
      + user_data_base64                     = (known after apply)
      + user_data_replace_on_change          = false
      + vpc_security_group_ids               = (known after apply)

      + ebs_block_device {
          + delete_on_termination = true
          + device_name           = "my_data"
          + encrypted             = (known after apply)
          + iops                  = 1200
          + kms_key_id            = (known after apply)
          + snapshot_id           = (known after apply)
          + tags_all              = (known after apply)
          + throughput            = (known after apply)
          + volume_id             = (known after apply)
          + volume_size           = 3000
          + volume_type           = "io1"
        }

      + root_block_device {
          + delete_on_termination = true
          + device_name           = (known after apply)
          + encrypted             = (known after apply)
          + iops                  = (known after apply)
          + kms_key_id            = (known after apply)
          + tags_all              = (known after apply)
          + throughput            = (known after apply)
          + volume_id             = (known after apply)
          + volume_size           = 50
          + volume_type           = (known after apply)
        }
    }

Plan: 1 to add, 0 to change, 0 to destroy.
╷
│ Warning: AWS account ID not found for provider
│ 
│   with provider["registry.terraform.io/hashicorp/aws"],
│   on main.tf line 1, in provider "aws":
│    1: provider "aws" {
│ 
│ See
│ https://registry.terraform.io/providers/hashicorp/aws/latest/docs#skip_requesting_account_id
│ for implications.
╵

2024-04-10T17:28:29Z INF Detected Terraform plan JSON file at /root/.atlantis/repos/infracost/atlantis-demo/23/default/terraform-aws/default.json
2024-04-10T17:28:29Z INF Starting: Extracting only cost-related params from terraform
2024-04-10T17:28:29Z INF Starting: Retrieving cloud prices to calculate costs
2024-04-10T17:28:29Z INF Output saved to /tmp/infracost-atlantis-demo-23/default-terraform-aws-infracost.json
total 8K     
drwxr-xr-x    2 root     root          50 Apr 10 17:28 .
drwxrwxrwt    1 root     root          40 Apr 10 17:28 ..
-rw-r--r--    1 root     root        4.2K Apr 10 17:28 default-terraform-aws-infracost.json
```

* :arrow_forward: To **apply** this plan, comment:
    * `atlantis apply -d terraform-aws`
* :put_litter_in_its_place: To **delete** this plan click [here](http://45e0648578db:4141/lock?id=infracost%252Fatlantis-demo%252Fterraform-aws%252Fdefault)
* :repeat: To **plan** this project again, comment:
    * `atlantis plan -d terraform-aws`
</details>
Plan: 1 to add, 0 to change, 0 to destroy.

---
### 2. dir: `terraform-gcp` workspace: `default`
<details><summary>Show Output</summary>

```diff
Terraform used the selected providers to generate the following execution
plan. Resource actions are indicated with the following symbols:
+ create

Terraform will perform the following actions:

  # google_compute_instance.instance1 will be created
+ resource "google_compute_instance" "instance1" {
      + can_ip_forward       = false
      + cpu_platform         = (known after apply)
      + current_status       = (known after apply)
      + deletion_protection  = false
      + effective_labels     = (known after apply)
      + guest_accelerator    = [
          + {
              + count = 4
              + type  = "nvidia-tesla-t4"
            },
        ]
      + id                   = (known after apply)
      + instance_id          = (known after apply)
      + label_fingerprint    = (known after apply)
      + machine_type         = "n1-standard-32"
      + metadata_fingerprint = (known after apply)
      + min_cpu_platform     = (known after apply)
      + name                 = "instance1"
      + project              = "my-project"
      + self_link            = (known after apply)
      + tags_fingerprint     = (known after apply)
      + terraform_labels     = (known after apply)
      + zone                 = "us-central1-a"

      + boot_disk {
          + auto_delete                = true
          + device_name                = (known after apply)
          + disk_encryption_key_sha256 = (known after apply)
          + kms_key_self_link          = (known after apply)
          + mode                       = "READ_WRITE"
          + source                     = (known after apply)

          + initialize_params {
              + image                  = "debian-cloud/debian-9"
              + labels                 = (known after apply)
              + provisioned_iops       = (known after apply)
              + provisioned_throughput = (known after apply)
              + size                   = (known after apply)
              + type                   = (known after apply)
            }
        }

      + network_interface {
          + internal_ipv6_prefix_length = (known after apply)
          + ipv6_access_type            = (known after apply)
          + ipv6_address                = (known after apply)
          + name                        = (known after apply)
          + network                     = "default"
          + network_ip                  = (known after apply)
          + stack_type                  = (known after apply)
          + subnetwork                  = (known after apply)
          + subnetwork_project          = (known after apply)

          + access_config {
              + nat_ip       = (known after apply)
              + network_tier = (known after apply)
            }
        }

      + scheduling {
          + automatic_restart   = true
          + on_host_maintenance = (known after apply)
          + preemptible         = true
          + provisioning_model  = (known after apply)
        }
    }

  # google_dns_record_set.frontend will be created
+ resource "google_dns_record_set" "frontend" {
      + id           = (known after apply)
      + managed_zone = "zone"
      + name         = "frontend.123."
      + project      = "my-project"
      + rrdatas      = [
          + "123.123.123.123]",
        ]
      + ttl          = 300
      + type         = "A"
    }

Plan: 2 to add, 0 to change, 0 to destroy.


2024-04-10T17:28:33Z INF Detected Terraform plan JSON file at /root/.atlantis/repos/infracost/atlantis-demo/23/default/terraform-gcp/default.json
2024-04-10T17:28:33Z INF Starting: Extracting only cost-related params from terraform
2024-04-10T17:28:33Z INF Starting: Retrieving cloud prices to calculate costs
2024-04-10T17:28:33Z INF Output saved to /tmp/infracost-atlantis-demo-23/default-terraform-gcp-infracost.json
total 12K    
drwxr-xr-x    2 root     root          94 Apr 10 17:28 .
drwxrwxrwt    1 root     root          40 Apr 10 17:28 ..
-rw-r--r--    1 root     root        4.2K Apr 10 17:28 default-terraform-aws-infracost.json
-rw-r--r--    1 root     root        3.9K Apr 10 17:28 default-terraform-gcp-infracost.json
```

* :arrow_forward: To **apply** this plan, comment:
    * `atlantis apply -d terraform-gcp`
* :put_litter_in_its_place: To **delete** this plan click [here](http://45e0648578db:4141/lock?id=infracost%252Fatlantis-demo%252Fterraform-gcp%252Fdefault)
* :repeat: To **plan** this project again, comment:
    * `atlantis plan -d terraform-gcp`
</details>
Plan: 2 to add, 0 to change, 0 to destroy.

---
* :fast_forward: To **apply** all unapplied plans from this pull request, comment:
    * `atlantis apply`
* :put_litter_in_its_place: To **delete** all plans and locks for the PR, comment:
    * `atlantis unlock`